+++ src/main/java/com/dougfsilva/e_AGE/application/usecases/course/DeleteCourse.java
@@ -2,8 +2,10 @@ package com.dougfsilva.e_AGE.application.usecases.course;
 
 import com.dougfsilva.e_AGE.application.usecases.utilities.DeleteImage;
 import com.dougfsilva.e_AGE.application.usecases.utilities.StandardLogger;
+import com.dougfsilva.e_AGE.domain.clazz.ClazzRepository;
 import com.dougfsilva.e_AGE.domain.course.Course;
 import com.dougfsilva.e_AGE.domain.course.CourseRepository;
+import com.dougfsilva.e_AGE.domain.utilities.exception.DataIntegrityViolationException;
 import com.dougfsilva.e_AGE.domain.utilities.image.ImageType;
 
 import lombok.AllArgsConstructor;
@@ -13,6 +15,8 @@ public class DeleteCourse {
 
 	private final CourseRepository repository;
 	
+	private final ClazzRepository clazzRepository;
+	
 	private final FindCourse findCourse;
 	
 	private final DeleteImage deleteImage;
@@ -21,6 +25,11 @@ public class DeleteCourse {
 	
 	public void execute(String ID) {
 		Course course = findCourse.findByID(ID);
+		if(clazzRepository.countByCourse(course) > 0) {
+			throw new DataIntegrityViolationException(
+					String.format("The Course %S cannot be deleted because there are classes still associated with it", 
+							course.getTitle()));
+		}
 		repository.delete(course);
 		deleteImage.execute(course.getImage(), ImageType.COURSE);
 		logger.deletedObjectLog(course);
