+++ src/main/java/com/dougfsilva/e_AGE/application/usecases/technologicalArea/DeleteTechnologicalArea.java
@@ -2,8 +2,10 @@ package com.dougfsilva.e_AGE.application.usecases.technologicalArea;
 
 import com.dougfsilva.e_AGE.application.usecases.utilities.DeleteImage;
 import com.dougfsilva.e_AGE.application.usecases.utilities.StandardLogger;
+import com.dougfsilva.e_AGE.domain.course.CourseRepository;
 import com.dougfsilva.e_AGE.domain.technologicalArea.TechnologicalArea;
 import com.dougfsilva.e_AGE.domain.technologicalArea.TechnologicalAreaRepository;
+import com.dougfsilva.e_AGE.domain.utilities.exception.DataIntegrityViolationException;
 import com.dougfsilva.e_AGE.domain.utilities.image.ImageType;
 
 import lombok.AllArgsConstructor;
@@ -13,6 +15,8 @@ public class DeleteTechnologicalArea {
 
 	private final TechnologicalAreaRepository repository;
 	
+	private final CourseRepository courseRepository;
+	
 	private final FindTechnologicalArea findTechnologicalArea;
 	
 	private final DeleteImage deleteImage ;
@@ -21,6 +25,11 @@ public class DeleteTechnologicalArea {
 	
 	public void delete(String ID) {
 		TechnologicalArea technologicalArea = findTechnologicalArea.findByID(ID);
+		if (courseRepository.countByTechnologialArea(technologicalArea) > 0) {
+			throw new DataIntegrityViolationException(
+					String.format("The technological area %S cannot be deleted because there are courses still associated with it", 
+							technologicalArea.getTitle()));
+		}
 		repository.delete(technologicalArea);
 		deleteImage.execute(ID, ImageType.TECHNOLOGICAL_AREA);
 		logger.deletedObjectLog(technologicalArea);
